Move EscapeSequences into base
