+++ src/InputHandler.ts
@@ -5,12 +5,13 @@
  */
 
 import { CharData, IInputHandler, IDcsHandler, IEscapeSequenceParser, IBuffer } from './Types';
-import { C0, C1 } from './EscapeSequences';
+import { C0, C1 } from './base/data/EscapeSequences';
 import { CHARSETS, DEFAULT_CHARSET } from './core/data/Charsets';
 import { CHAR_DATA_CHAR_INDEX, CHAR_DATA_WIDTH_INDEX, CHAR_DATA_CODE_INDEX, DEFAULT_ATTR } from './Buffer';
 import { FLAGS } from './renderer/Types';
 import { wcwidth } from './CharWidth';
 import { EscapeSequenceParser } from './EscapeSequenceParser';
+import { ICharset } from './core/Types';
 
 /**
  * Map collect to glevel. Used in `selectCharset`.
