+++ src/base/data/EscapeSequences.ts
