+++ src/core/input/Keyboard.ts
@@ -6,7 +6,7 @@
 
 import { IKeyboardEvent } from '../../base/Types';
 import { IKeyboardResult, KeyboardResultType } from '../Types';
-import { C0 } from '../../EscapeSequences';
+import { C0 } from '../../base/data/EscapeSequences';
 
 // reg + shift key mappings for digits and special chars
 const KEYCODE_KEY_MAPPINGS: { [key: number]: [string, string]} = {
